refactor(web): decompose workbench app composition
