+++ apps/web/src/features/workbench/components/workbench-layout.tsx
@@ -1,0 +1,22 @@
+import * as React from "react"
+
+import { SidebarInset, SidebarProvider, SidebarRail } from "../../../components/ui/sidebar"
+
+type WorkbenchLayoutProps = {
+  sidebar: React.ReactNode
+  header: React.ReactNode
+  children: React.ReactNode
+}
+
+export const WorkbenchLayout = ({ sidebar, header, children }: WorkbenchLayoutProps) => {
+  return (
+    <SidebarProvider>
+      {sidebar}
+      <SidebarRail />
+      <SidebarInset>
+        {header}
+        {children}
+      </SidebarInset>
+    </SidebarProvider>
+  )
+}
